Click to Return Button is usable
Click to Return Button will (depending on what is commented in or out) delete the row clicked or keep the row and change the button to a time stamp. A bug is that clicking anywhere in the row will also count as clicking the button, and the timestamps can be overwritten by clicking anywhere in the row.